Rezept löschen

+++ grails-app/views/recipe/list.gsp
@@ -35,7 +35,7 @@
                     <td>${rec.category}</td>
                     <td class="text-right">
                         <a class="btn btn-outline-secondary btn-sm" href="/recipe/edit?id=${rec.id}"><i class="fa fa-pencil"></i></a>
-                        <a class="btn btn-outline-danger btn-sm" href="/recipe/delete?id=${rec.id}"><i class="fa fa-trash"></i></a>
+                        <a class="btn btn-outline-danger btn-sm" href="/recipe/delete?id=${rec.id}" onclick="return confirm('Wirklich löschen?');"><i class="fa fa-trash"></i></a>
                     </td>
                 </tr>
             </g:each>
